adding random_normal_like layer

+++ test/layer_multiply.cc
@@ -8,7 +8,7 @@ int main()
     auto a = variable{ ones<float>( {2, 2} ) };
     auto b = variable{ ones<float>( {2, 2} ) };
     auto ab = a+b; // or 'auto ab = multiply( a, b );'
-    ceras::session<ceras::tensor<double>> s;
+    ceras::session<ceras::tensor<float>> s;
     std::cout <<  s.run( ab );
 
     return 0;
